Fixing error in evics-post-command-hook
This error was causing the function to be removed from the post-command-hook.
Turns out, the bug was related to deleting text when the cursor was at the
end of the file. The subsequent invocation of evics-pch would try to translate
a point that is out of bounds to a line number.